+++ mycli/main.py
@@ -580,10 +580,7 @@ class MyCli(object):
                 try:
                     special.write_tee('\n'.join(output))
                     special.write_once('\n'.join(output))
-                    if special.is_pager_enabled():
-                        self.output_via_pager('\n'.join(output))
-                    else:
-                        self.output('\n'.join(output))
+                    self.output('\n'.join(output))
                 except KeyboardInterrupt:
                     pass
                 if special.is_timing_enabled():
@@ -643,18 +640,26 @@ class MyCli(object):
             if not self.less_chatty:
                 self.output('Goodbye!')
 
-    def output(self, text, **kwargs):
-        special.write_tee(text)
+    def output(self, output, **kwargs):
+        special.write_tee(output)
         if self.logfile:
-            self.logfile.write(utf8tounicode(text))
+            self.logfile.write(utf8tounicode(output))
             self.logfile.write('\n')
-        click.secho(text, **kwargs)
 
-    def output_via_pager(self, text):
-        if self.logfile:
-            self.logfile.write(text)
-            self.logfile.write('\n')
-        click.echo_via_pager(text)
+        size = self.cli.output.get_size()
+
+        pager = False
+        margin = self.get_reserved_space() + self.get_prompt(self.prompt_format).count('\n') + 1
+        for i, line in enumerate(output.splitlines()):
+            if len(line) > size.columns or i > (size.rows - margin):
+                pager = True
+                break
+
+        if (special.is_pager_enabled() and pager) or self.explicit_pager:
+            click.echo_via_pager(output)
+        else:
+            click.secho(output, **kwargs)
+
 
     def configure_pager(self):
         # Provide sane defaults for less if they are empty.
@@ -664,6 +669,10 @@ class MyCli(object):
         cnf = self.read_my_cnf_files(self.cnf_files, ['pager', 'skip-pager'])
         if cnf['pager']:
             special.set_pager(cnf['pager'])
+            self.explicit_pager = True
+        else:
+            self.explicit_pager = False
+
         if cnf['skip-pager']:
             special.disable_pager()